[Proposal] Include a `t` function
The t function helps localizing strings where the component cannot be
used. Some examples are:

- You need a list of localized strings alphabetically sorted.
- Want to use as text in an ion-select.
- Want to use as parameters of other functions (date formatting, as an
example).

It follows a few conventions I saw in many other libraries through the
years, while leaving others I'm not sure how I want them implemented
yet.

The main one is using the parameter `count` in replacements as the way
to pluralize the string. If a `count` field is found, we look for
a pluralize suffix. That is, to the key being used, we'll append the
result of running the `select` method to the `count` parameter. (Getting
the `locale` can certainly be improved, and maybe would need to be moved
to dictionary.